requirements.txt
# Python 3.9.12
# torch==1.13.1+cu117
# torchvision==0.14.1+cu117
# pip install torch==1.13.1+cu117 torchvision==0.14.1+cu117 --extra-index-url https://download.pytorch.org/whl/cu117

pytorch-lightning==1.8.4
torchmetrics==0.11.0
torch-geometric==2.3.1
torch-scatter==2.1.1

numpy==1.24.3
scipy==1.13.1
pandas==2.3.2
scikit-learn==1.2.2
statsmodels==0.14.5

matplotlib==3.7.1
seaborn==0.12.2
tueplots==0.0.5
networkx==2.8.8
igraph==0.11.9

wandb==0.15.8
yacs==0.1.8
pyyaml==6.0.2
tqdm==4.67.1
pytest==7.2.0
autograd==1.8.0

requests==2.32.4
joblib==1.5.1
fonttools==4.59.0
kiwisolver==1.4.7
cycler==0.12.1
pyparsing==3.2.3
contourpy==1.3.0
pillow==11.3.0
attrs==25.3.0
typing-extensions==4.14.1
dcor==0.6
pygam==0.10.1
kneed==0.8.5

## gpr
cupy==13.4.1